Fix statics refactor to command and at least builds. Fails on javaDoc build.
