+++ stream/src/main/java/module-info.java
@@ -77,7 +77,6 @@ module com.hedera.block.stream {
     requires transitive io.grpc;
     requires io.grpc.protobuf;
     requires org.antlr.antlr4.runtime;
-    requires com.google.cloud.core;
     requires static com.github.spotbugs.annotations;
     requires static java.annotation;
 }
