+++ tools/src/main/java/com/hedera/block/tools/commands/record2blocks/gcp/MainNetBucket.java
@@ -1,9 +1,5 @@
 package com.hedera.block.tools.commands.record2blocks.gcp;
 
-import static com.hedera.block.tools.commands.record2blocks.Main.DATA_DIR;
-import static com.hedera.block.tools.commands.record2blocks.Main.MAX_NODE_ACCOUNT_ID;
-import static com.hedera.block.tools.commands.record2blocks.Main.MIN_NODE_ACCOUNT_ID;
-
 import com.google.cloud.storage.Bucket;
 import com.google.cloud.storage.Storage;
 import com.google.cloud.storage.Storage.BlobField;
@@ -17,7 +13,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
@@ -36,14 +31,6 @@ import java.util.zip.GZIPOutputStream;
  * </p>
  */
 public class MainNetBucket {
-    /**
-     * The cache enabled switch. When caching is enabled all fetched data is saved on disk and reused between runs. This
-     * is useful for debugging and development. But when we get to doing long runs with many TB of data this is not
-     * practical.
-     */
-    public static final AtomicBoolean CACHE_ENABLED = new AtomicBoolean(true);
-    /** The cache directory, where we store all downloaded content for reuse if CACHE_ENABLED is true. */
-    public static final Path CACHE_DIR = DATA_DIR.resolve("gcp-cache");
     /** The required fields we need from blobs */
     private static final Storage.BlobListOption REQUIRED_FIELDS =
             BlobListOption.fields(BlobField.NAME, BlobField.SIZE, BlobField.MD5HASH);
@@ -54,21 +41,51 @@ public class MainNetBucket {
             .get(HEDERA_MAINNET_STREAMS_BUCKET);
 
     /**
+     * The cache enabled switch. When caching is enabled all fetched data is saved on disk and reused between runs. This
+     * is useful for debugging and development. But when we get to doing long runs with many TB of data this is not
+     * practical.
+     */
+    private final boolean cacheEnabled;
+
+    /** The cache directory, where we store all downloaded content for reuse if CACHE_ENABLED is true. */
+    private final Path cacheDir;// = DATA_DIR.resolve("gcp-cache");
+
+    /** The minimum node account id in the network. */
+    private final int minNodeAccountId;
+
+    /** The maximum node account id in the network. */
+    private final int maxNodeAccountId;
+    /**
+     * Create a new MainNetBucket instance with the given cache enabled switch and cache directory.
+     *
+     * @param cacheEnabled the cache enabled switch
+     * @param cacheDir the cache directory
+     * @param minNodeAccountId the minimum node account id in the network
+     * @param maxNodeAccountId the maximum node account id in the network
+     */
+    public MainNetBucket(boolean cacheEnabled, Path cacheDir, int minNodeAccountId, int maxNodeAccountId) {
+        this.cacheEnabled = cacheEnabled;
+        this.cacheDir = cacheDir;
+        this.minNodeAccountId = minNodeAccountId;
+        this.maxNodeAccountId = maxNodeAccountId;
+    }
+
+    /**
      * Download a file from GCP, caching if CACHE_ENABLED is true. This is designed to be thread safe.
      *
      * @param path the path to the file in the bucket
      * @return the bytes of the file
      */
-    public static byte[] download(String path) {
+    public byte[] download(String path) {
         try {
-            Path cachedFilePath = CACHE_DIR.resolve(path);
-            if (CACHE_ENABLED.get() && Files.exists(cachedFilePath)) {
+            Path cachedFilePath = cacheDir.resolve(path);
+            if (cacheEnabled && Files.exists(cachedFilePath)) {
                 return Files.readAllBytes(cachedFilePath);
             } else {
                 byte[] bytes = STREAMS_BUCKET.get(path).getContent();
-                if (CACHE_ENABLED.get()) {
+                if (cacheEnabled) {
                     Files.createDirectories(cachedFilePath.getParent());
-                    Path tempCachedFilePath = Files.createTempFile(CACHE_DIR, null, ".tmp");
+                    Path tempCachedFilePath = Files.createTempFile(cacheDir, null, ".tmp");
                     Files.write(tempCachedFilePath, bytes);
                     Files.move(tempCachedFilePath, cachedFilePath);
                 }
@@ -85,7 +102,7 @@ public class MainNetBucket {
      * @param blockStartTime the start time of a block, in nanoseconds since OA
      * @return a stream of ChainFiles that contain the records for the given day.
      */
-    public static List<ChainFile> listDay(long blockStartTime) {
+    public List<ChainFile> listDay(long blockStartTime) {
         final String datePrefix = RecordFileDates.blockTimeLongToRecordFilePrefix(blockStartTime);
         // crop to the hour
         final String dayPrefix = datePrefix.substring(0, datePrefix.indexOf('T'));
@@ -100,7 +117,7 @@ public class MainNetBucket {
      * @param blockStartTime the start time of a block, in nanoseconds since OA
      * @return a stream of ChainFiles that contain the records for the given hour.
      */
-    public static List<ChainFile> listHour(long blockStartTime) {
+    public List<ChainFile> listHour(long blockStartTime) {
         final String datePrefix = RecordFileDates.blockTimeLongToRecordFilePrefix(blockStartTime);
         // crop to the hour
         final String hourPrefix = datePrefix.substring(0, datePrefix.indexOf('_'));
@@ -114,11 +131,11 @@ public class MainNetBucket {
      * @param filePrefix the prefix of the file name to search for
      * @return a stream of ChainFiles that have a filename that starts with the given prefix.
      */
-    private static List<ChainFile> listWithFilePrefix(String filePrefix) {
+    private List<ChainFile> listWithFilePrefix(String filePrefix) {
         try {
             // read from cache if it already exists in cache
-            final Path listCacheFilePath = CACHE_DIR.resolve("list-"+filePrefix+".bin.gz");
-            if (CACHE_ENABLED.get() && Files.exists(listCacheFilePath)) {
+            final Path listCacheFilePath = cacheDir.resolve("list-"+filePrefix+".bin.gz");
+            if (cacheEnabled && Files.exists(listCacheFilePath)) {
                 try(ObjectInputStream ois = new ObjectInputStream(
                         new GZIPInputStream(Files.newInputStream(listCacheFilePath)))) {
                     final int fileCount = ois.readInt();
@@ -130,7 +147,7 @@ public class MainNetBucket {
                 }
             }
             // create a list of ChainFiles
-            List<ChainFile> chainFiles = IntStream.range(MIN_NODE_ACCOUNT_ID, MAX_NODE_ACCOUNT_ID + 1)
+            List<ChainFile> chainFiles = IntStream.range(minNodeAccountId, maxNodeAccountId + 1)
                     .parallel()
                     .mapToObj(nodeAccountId ->
                             Stream.concat(
@@ -148,7 +165,7 @@ public class MainNetBucket {
                     ).flatMap(Function.identity())
                     .toList();
             // save the list to cache
-            if (CACHE_ENABLED.get()) {
+            if (cacheEnabled) {
                 Files.createDirectories(listCacheFilePath.getParent());
                 try(ObjectOutputStream oos = new ObjectOutputStream(
                         new GZIPOutputStream(Files.newOutputStream(listCacheFilePath)))) {
@@ -165,20 +182,20 @@ public class MainNetBucket {
         }
     }
 
-    /**
-     * Main method to test the listing of blobs for a given hour.
-     *
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        long oneHourInNanos = 3_600_000_000_000L;
-        long hours = Long.MAX_VALUE / oneHourInNanos;
-        System.out.println("hours = " + hours);
-        long days = hours / 24;
-        System.out.println("days = " + days);
-        long years = days / 365;
-        System.out.println("years = " + years);
-//        listHour(oneHourInNanos*24*365*5).forEach(chainFile -> {
-        listHour(0).forEach(System.out::println);
-    }
+//    /**
+//     * Main method to test the listing of blobs for a given hour.
+//     *
+//     * @param args the command line arguments
+//     */
+//    public static void main(String[] args) {
+//        long oneHourInNanos = 3_600_000_000_000L;
+//        long hours = Long.MAX_VALUE / oneHourInNanos;
+//        System.out.println("hours = " + hours);
+//        long days = hours / 24;
+//        System.out.println("days = " + days);
+//        long years = days / 365;
+//        System.out.println("years = " + years);
+////        listHour(oneHourInNanos*24*365*5).forEach(chainFile -> {
+//        listHour(0).forEach(System.out::println);
+//    }
 }
