+++ tools/src/main/java/com/hedera/block/tools/commands/record2blocks/mirrornode/RecordFileCsvExtractBlockTimes.java
@@ -1,7 +1,5 @@
 package com.hedera.block.tools.commands.record2blocks.mirrornode;
 
-import static com.hedera.block.tools.commands.record2blocks.Main.DATA_DIR;
-import static com.hedera.block.tools.commands.record2blocks.mirrornode.MirrorNodeConstants.RECORDS_CSV_FILE;
 import static com.hedera.block.tools.commands.record2blocks.util.RecordFileDates.recordFileNameToBlockTimeLong;
 
 import java.io.BufferedReader;
@@ -14,22 +12,38 @@ import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.zip.GZIPInputStream;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
 
 /**
- * Read the record_file.csv.gz file from mirror node and count the number of lines in the file.
+ * Read the record_file.csv.gz file from mirror node and extract the block times into a file.
+ * <p>
+ * The block times file is a binary file of longs, each long is the number of nanoseconds for that block after first
+ * block time. So first block = 0, second about 5 seconds later etc. The index is the block number, so block 0 is first
+ * long, block 1 is second block and so on.
+ * </p>
  */
 @SuppressWarnings({"DuplicatedCode", "CallToPrintStackTrace"})
-public class RecordFileCsvExtractBlockTimes {
-    /**
-     * The path to the block times file, this is a binary file of longs, each long is the number of nanoseconds for
-     * that block after first block time. So first block = 0, second about 5 seconds later etc. The index is the block
-     * number, so block 0 is first long, block 1 is second block and so on.
-     */
-    public static final Path BLOCK_TIMES_FILE = DATA_DIR.resolve("block_times.bin");
-    /** the number of blocks in the record CSV file roughly */
-    public static int NUMBER_OF_BLOCKS_ROUNDED_UP = 70_000_000;
+@Command(name = "extractBlockTimes", description = "Extract block times from mirror node records csv file")
+public class RecordFileCsvExtractBlockTimes implements Runnable {
+    /** the number of blocks in the record CSV file roughly, used for progress estimation */
+    private static final int NUMBER_OF_BLOCKS_ROUNDED_UP = 70_000_000;
 
-    public static void main(String[] args) {
+    /** The path to the record table CSV from mirror node, gzipped. */
+    @Option(names = {"--record-csv"},
+            description = "Path to the record table CSV from mirror node, gzipped.")
+    private Path recordsCsvFile = Path.of("data/record_file.csv.gz");
+
+    /** The path to the block times file. */
+    @Option(names = {"--block-times"},
+            description = "Path to the block times \".bin\" file.")
+    private Path blockTimesFile = Path.of("data/block_times.bin");
+
+    /**
+     * Read the record file table CSV file and extract the block times into a file.
+     */
+    @Override
+    public void run() {
         // get the start time of the first block
         // create off heap array to store the block times
         final ByteBuffer blockTimesBytes = ByteBuffer.allocateDirect(NUMBER_OF_BLOCKS_ROUNDED_UP * Long.BYTES);
@@ -38,7 +52,7 @@ public class RecordFileCsvExtractBlockTimes {
         final AtomicInteger blockCount = new AtomicInteger(0);
         // read the record file table CSV file
         try(var reader = new BufferedReader(new InputStreamReader(new GZIPInputStream(
-                new FileInputStream(RECORDS_CSV_FILE.toFile()))))) {
+                new FileInputStream(recordsCsvFile.toFile()))))) {
             // skip header
             reader.readLine();
             // read all lines
@@ -73,7 +87,7 @@ public class RecordFileCsvExtractBlockTimes {
             }
             System.out.println("\ntotalBlocksWithoutTimes = " + totalBlocksWithoutTimes);
             // write the block times to a file
-            try (final var out = Files.newByteChannel(BLOCK_TIMES_FILE, StandardOpenOption.CREATE, StandardOpenOption.WRITE)) {
+            try (final var out = Files.newByteChannel(blockTimesFile, StandardOpenOption.CREATE, StandardOpenOption.WRITE)) {
                 blockTimesBytes.position(0);
                 long bytesWritten = out.write(blockTimesBytes);
                 System.out.println("bytesWritten = " + bytesWritten);
